+++ src/cli/test.rs
@@ -40,9 +40,13 @@ pub struct Args {
     #[arg(long, default_value_t = crate::cli::default_threads(), value_name = "THREADS")]
     pub threads: usize,
 
-    /// Maximum number of sequences to run across all threads.
-    #[arg(long, default_value_t = 256, value_name = "RUNS")]
-    pub max_runs: u64,
+    /// Maximum number of sequences to fuzz, split across all threads.
+    #[arg(long, default_value_t = 100_000, value_name = "RUNS")]
+    pub max_fuzz_runs: u64,
+
+    /// Maximum number of shrink attempts, split across all threads.
+    #[arg(long, default_value_t = 10_000, value_name = "RUNS")]
+    pub max_shrink_runs: u64,
 
     /// Maximum number of handler calls per sequence.
     #[arg(long, default_value_t = 8, value_name = "COUNT")]
@@ -202,7 +206,7 @@ pub fn run(args: Args) -> Result<Vec<BrokenInvariant>> {
         .with_coverage(coverage.clone())
         .with_broken_invariants(shared_broken_invariants)
         .with_threads(args.threads)
-        .with_max_runs(args.max_runs)
+        .with_max_runs(args.max_fuzz_runs)
         .with_max_calls(args.max_calls)
         .with_timeout(args.timeout.map(Duration::from_secs))
         .with_seed(seed);
@@ -226,7 +230,7 @@ pub fn run(args: Args) -> Result<Vec<BrokenInvariant>> {
             .with_target(address)
             .with_deployer(deployer)
             .with_threads(args.threads)
-            .with_max_runs(args.max_runs)
+            .with_max_runs(args.max_shrink_runs)
             .with_timeout(args.timeout.map(Duration::from_secs))
             .with_seed(seed)
             .shrink(&broken_invariants)?;